Refactor: separate tests from fixture.
