+++ tests/run_tests.py
@@ -4,10 +4,8 @@ import argparse
 import re
 from shutil import rmtree, copytree
 from ci_utils import xqt, pushd
+from utils import COVER_DIRS
 
-COVER_DIRS = 'applications/runestone/tests,applications/runestone/controllers,applications/runestone/models'
-
-# Assume we are running with working directory in tests
 
 if __name__ == '__main__':
     os.environ['WEB2PY_CONFIG'] = 'test'
@@ -59,9 +57,9 @@ if __name__ == '__main__':
         print('Skipping DB initialization.')
     else:
         # make sure runestone_test is nice and clean.
-        xqt('dropdb --echo --if-exists --host={} --user={} "{}"'.format(pgnetloc, pguser, dbname),
-            'createdb --echo --host={} --user={} "{}"'.format(pgnetloc, pguser, dbname),
-            'psql  --host={} --user={} "{}" < runestone_test.sql'.format(pgnetloc, pguser, dbname))
+        xqt('dropdb --echo --if-exists --host={} --username={} "{}"'.format(pgnetloc, pguser, dbname),
+            'createdb --echo --host={} --username={} "{}"'.format(pgnetloc, pguser, dbname),
+            'psql  --host={} --username={} "{}" < runestone_test.sql'.format(pgnetloc, pguser, dbname))
         # Copy the test book to the books directory.
         rmtree('../books/test_course_1', ignore_errors=True)
         # Sometimes this fails for no good reason on Windows. Retry.
@@ -87,5 +85,5 @@ if __name__ == '__main__':
             '{} -m pytest -v applications/runestone/tests/test_server.py {}'.format(sys.executable, ' '.join(extra_args)),
             *['{} -m coverage run --append --source={} web2py.py -S runestone -M -R applications/runestone/tests/{}'.format(sys.executable, COVER_DIRS, x)
               for x in ['test_ajax.py', 'test_dashboard.py', 'test_admin.py', 'test_assignments.py']]
-            )
+        )
         xqt('{} -m coverage report'.format(sys.executable))
